backend/requirements.txt
boto3==1.26.157
botocore==1.29.157
Flask==2.3.2
Flask_Cors==3.0.10
numpy==1.24.3
opencv_python==4.7.0.72
python-dotenv==1.0.0
replicate==0.8.1
Requests==2.31.0
